+++ odoo_backup/__init__.py
@@ -1,2 +1,2 @@
 """Odoo Backup Tool Package"""
-__version__ = "0.2.1"
+__version__ = "0.2.2"